ai-model/requirements.txt
mkdir C:\tmp
$env:TEMP='C:\tmp'; $env:TMP='C:\tmp'
python -m pip install --upgrade pip setuptools wheel
python -m pip install -r ai-model\requirements.txt# Data Processing
numpy>=1.26.0
pandas>=2.1.0

# ML/DL Framework
torch>=2.1.0
scikit-learn>=1.4.0

# Technical Indicators
ta>=0.11.0
pandas-ta>=0.3.14b

# Binance Data Collection
python-binance>=1.0.19
ccxt>=4.2.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
joblib>=1.3.0
